backend/requirements.txt
fastapi
uvicorn
websockets
google-cloud-aiplatform
gunicorn